refactor - Mudança dos Id's de Int para Long

+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/domain/atividades/Ano.kt
@@ -9,13 +9,13 @@ import jakarta.validation.constraints.NotBlank
 @Entity
 class Ano(
     @field:Id
-    @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idAno: Int = 0,
+    @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idAno: Long = 0,
     @field:NotBlank @field:Max(60) private var ano:String
 ) {
-    fun getId():Int{
+    fun getId():Long{
         return idAno
     }
-    fun setId(novoId:Int){
+    fun setId(novoId:Long){
         idAno = novoId
     }
     fun getAno():String{
